+++ src/app/organizations/policies/personal-ownership.component.ts
@@ -4,7 +4,7 @@ import { BasePolicy } from 'jslib-common/abstractions/policy.service';
 
 import { PolicyType } from 'jslib-common/enums/policyType';
 
-import { BasePolicyComponent } from '../manage/policies.component';
+import { BasePolicyComponent } from '../manage/policy-edit.component';
 
 export class PersonalOwnershipPolicy extends BasePolicy {
     name = 'personalOwnership';
