+++ src/app/organizations/policies/require-sso.component.ts
@@ -8,7 +8,7 @@ import { PolicyType } from 'jslib-common/enums/policyType';
 import { Organization } from 'jslib-common/models/domain/organization';
 import { PolicyRequest } from 'jslib-common/models/request/policyRequest';
 
-import { BasePolicyComponent } from '../manage/policies.component';
+import { BasePolicyComponent } from '../manage/policy-edit.component';
 
 export class RequireSsoPolicy extends BasePolicy {
     name = 'requireSso';
