+++ src/app/organizations/policies/reset-password.component.ts
@@ -7,7 +7,7 @@ import { PolicyType } from 'jslib-common/enums/policyType';
 
 import { Organization } from 'jslib-common/models/domain/organization';
 
-import { BasePolicyComponent } from '../manage/policies.component';
+import { BasePolicyComponent } from '../manage/policy-edit.component';
 
 export class ResetPasswordPolicy extends BasePolicy {
     name = 'resetPasswordPolicy';
