+++ src/app/organizations/policies/send-options.component.ts
@@ -5,7 +5,7 @@ import { BasePolicy } from 'jslib-common/abstractions/policy.service';
 
 import { PolicyType } from 'jslib-common/enums/policyType';
 
-import { BasePolicyComponent } from '../manage/policies.component';
+import { BasePolicyComponent } from '../manage/policy-edit.component';
 
 export class SendOptionsPolicy extends BasePolicy {
     name = 'sendOptions';
